Extract StationRow; Array.includes
